+++ src/com/javaGelisme/Phone.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 public class Phone extends Products{
 
-    public Phone(int unitPrice, int discountRate, int amountOfStock,
+    public Phone(Double unitPrice, Double discountRate, int amountOfStock,
                  String productName,String productType, int id, Marka marka) {
         super(unitPrice, discountRate, amountOfStock, productName, productType, id, marka,
                 "64 GB - 128 GB", "Siyah-Kırmızı-Mavi", "6.1 Inc",
